Fixed player movement with dt Fixed sound

+++ Game/Source/Player.h
@@ -43,8 +43,8 @@ private:
 	bool wallSlideRight, wallSlideLeft;
 	bool stairs;
 
-	int remainingJumpSteps;
-	int remainingDash;
+	float remainingJumpSteps;
+	float remainingDash;
 	int dashCD;
 	bool isAttacking;
